refactor(backend): separate folder structure into domain and api layers
Reorganize backend codebase into a layered architecture:

- api/: HTTP-related code (routes, middleware, auth, config, db connections)
- domain/: Business logic (services, repositories, models, errors, utils)

This separation improves maintainability and makes dependencies clearer.
All import paths updated accordingly.
