+++ backend/src/api/middleware/auth.middleware.ts
