+++ backend/src/domain/models/external-identity.schema.ts
