+++ backend/src/domain/models/note.schema.ts
