+++ backend/src/domain/models/search.schema.ts
