+++ backend/src/domain/services/note/note.service.spec.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach } from 'vitest';
 // NOTE: Initialize database before importing db
-import '../../../tests/preload';
-import { db, notes, mentions } from '../../db';
+import '../../../../tests/preload';
+import { db, notes, mentions } from '../../../api/db';
 
 import { generateId } from '../../utils/id-generator';
 import { MAX_NOTE_LENGTH } from '@thread-note/shared/constants';
