+++ backend/src/domain/services/thread/thread.service.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach } from 'vitest';
-import '../../../tests/preload';
-import { db, notes, mentions } from '../../db';
+import '../../../../tests/preload';
+import { db, notes, mentions } from '../../../api/db';
 import { createThreadService } from '.';
 import { generateId } from '../../utils/id-generator';
