+++ backend/tests/integration/notes.routes.test.ts
@@ -10,8 +10,8 @@ import { env } from 'cloudflare:test';
 import { Hono } from 'hono';
 import type { ClerkClient } from '@clerk/backend';
 import notesRoutes from '../../src/api/routes/notes';
-import { errorHandler } from '../../src/middleware/error';
-import { generateId } from '../../src/utils/id-generator';
+import { errorHandler } from '../../src/api/middleware/error';
+import { generateId } from '../../src/domain/utils/id-generator';
 
 // NOTE: Mock the getClerkClient function
 const mockAuthenticateRequest = vi.fn();
